Linearization yaml, json schema, and sample linearization
A revised yaml linearization model, with backreference a uriorcurie, which is translated as a string in the generated json schema file. The label annotation property of Term/LanguageTerm was manually added, as the linkml schema generator apparently doesn't handle it. A sample linearization file contains two "LinearizationSection" backreferences and two linearization views (MMS and Dermatology) in each.  The sample was validated against the schema.